Refactor img load. Improve styling. Get img data.
Image loading from files is now handled by image models, and image models now
create and retain a reference to an HTML image element, which can be inserted
into the DOM when needed, and more importanly used to draw into a canvas
context.

To aid the latter, image models also generate and retain a reference to an
imageData array containing the RGBA values of the image pixels.

Various CSS and HTML improvements made, for make benefit glorious interface of
application.