Major refactoring to simplify and improve things.
Focus masks are now only generated when needed, and stored in the alpha channel
of each image, so toggling images on and off in Focus merge mode is
significantly faster.

Image models are now responsible for generating their focus masks, and do so in
a worker.

Estimation of time remaining has been improved.